- Agregación y modificación de métodos y variables

+++ src/main/java/com/abcm0018/inventarioautomatizado/users/service/impl/UserServiceImpl.java
@@ -5,6 +5,11 @@ import com.abcm0018.inventarioautomatizado.productos.domain.entity.Product;
 import com.abcm0018.inventarioautomatizado.productos.mappers.ProductMapper;
 import com.abcm0018.inventarioautomatizado.shared.config.InventariadoCacheConfig;
 import com.abcm0018.inventarioautomatizado.shared.utils.UserUtils;
+import com.abcm0018.inventarioautomatizado.timesheet.domain.entity.Timesheet;
+import com.abcm0018.inventarioautomatizado.timesheet.domain.repository.TimesheetRepository;
+import com.abcm0018.inventarioautomatizado.timesheet.dtos.TimesheetResponseDTO;
+import com.abcm0018.inventarioautomatizado.timesheet.mapper.TimesheetMapper;
+import com.abcm0018.inventarioautomatizado.users.domain.entity.Role;
 import com.abcm0018.inventarioautomatizado.users.domain.entity.User;
 import com.abcm0018.inventarioautomatizado.users.domain.repository.UserRepository;
 import com.abcm0018.inventarioautomatizado.users.dtos.UserRequest;
@@ -20,6 +25,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 @Transactional
@@ -28,6 +35,7 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final TimesheetRepository timesheetRepository;
 
     @Override
     @CacheEvict(allEntries = true)
@@ -38,12 +46,25 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByEmployeeNumber(employeeNumber).orElseThrow(() ->
                 new UserServiceException(CustomErrorCode.NOT_FOUND,
                         "The user with employee number " + employeeNumber + " doesn't exist",
-                        HttpStatus.BAD_REQUEST));
+                        HttpStatus.NOT_FOUND));
 
         User userToUpdate = UserMapper.toEntity(userRequest);
         userToUpdate.setId(user.getId());
         userToUpdate.setPassword(user.getPassword());
         userToUpdate.setRole(user.getRole());
+        userToUpdate.setRegistrationDate(user.getRegistrationDate());
+
+        if(userRequest.getActive() != null){
+            userToUpdate.setActive(userRequest.getActive());
+        }
+
+        if(userRequest.getExpirated() != null){
+            userToUpdate.setExpirated(userRequest.getExpirated());
+        }
+
+        if(userRequest.getBlocked() != null){
+            userToUpdate.setBlocked(userRequest.getBlocked());
+        }
 
         userRepository.save(userToUpdate);
 
@@ -59,7 +80,13 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByEmployeeNumber(employeeNumber).orElseThrow(() ->
                 new UserServiceException(CustomErrorCode.NOT_FOUND,
                         "The user with employee number " + employeeNumber + " doesn't exist",
-                        HttpStatus.BAD_REQUEST));
+                        HttpStatus.NOT_FOUND));
+
+        if(user.isActive()){
+            throw new UserServiceException(CustomErrorCode.CONFLICT,
+                    "The user with employee number " + employeeNumber + " cannot be deleted",
+                    HttpStatus.CONFLICT);
+        }
 
         userRepository.delete(user);
     }
@@ -73,13 +100,14 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByEmployeeNumber(employeeNumber).orElseThrow(() ->
                 new UserServiceException(CustomErrorCode.NOT_FOUND,
                         "The user with employee number " + employeeNumber + " doesn't exist",
-                        HttpStatus.BAD_REQUEST));
+                        HttpStatus.NOT_FOUND));
+
         return UserMapper.toDTO(user);
     }
 
     @Override
     public List<UserResponseDTO> getUsersByRole(String role) {
-        List<User> userList = userRepository.findByRole(role);
+        List<User> userList = userRepository.findByRole(Role.valueOf(role.toUpperCase()));
         return UserMapper.toDTOList(userList);
     }
